Some more ACO/ARO Work
Rearanged some pages to be easier to work with. Added a checkbox to
delete the field for an action in the aco_aro table when deleting the
action from the ActioNode Databse (because you don't always want to
delete actions from the aco_aro table when remvoing them from a
particular controller). Fixed a problem when attempting to create a
widget twice (when a widget was already in the acos_aros table). No
longer sved duplicated fields to the aco_aro table.